backend/requirements.txt
fastapi
uvicorn
requests
openai

aliyun-python-sdk-core
python-dotenv
